% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trocapipe.R
\name{trocapipe}
\alias{trocapipe}
\title{Troca pipe}
\usage{
trocapipe(file, from = "\%>\%", to = "|>", overwrite = FALSE)
}
\arguments{
\item{file}{String com o caminho ate o arquivo com o codigo.}

\item{from}{String indicando se eh o pipe tradicional "\%>\%" ou o
pipe novo "|>". O padrao eh o tradicional "\%>\%.}

\item{to}{String indicando se eh o pipe tradicional "\%>\%" ou o
pipe novo "|>". O padrao eh o novo "|>".}

\item{overwrite}{Se TRUE, sobresve o arquivo de entrada. Se FALSE
(padrao), coloca o codigo novo na tela.}
}
\value{
Retorna um TRUE invisivel caso ela funcione.
}
\description{
Troca o pipe tradicional pelo novo ou vice-versa.
}
\examples{

trocapipe(
  system.file("exemplos/exemplo_pipe_tradicional.R", package = "trocapipe")
)

\dontrun{
  trocapipe("file.R", "\%>\%", "|>")
}

}
