% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trocapipe.R
\name{trocapipe_tradicional}
\alias{trocapipe_tradicional}
\title{Troca pipe}
\usage{
trocapipe_tradicional(file)
}
\arguments{
\item{file}{String com o caminho ate o arquivo com o codigo.}
}
\value{
Retorna um TRUE invisivel caso ela funcione.
}
\description{
Troca o pipe tradicional pelo novo.
}
\examples{

\dontrun{
  trocapipe("file.R")
}

}
